Refactoring to remove utils package
