+++ pkg/apis/apis.go
@@ -1,6 +1,8 @@
 package apis
 
 import (
+	eventv1 "github.com/redhat-cop/events-notifier/pkg/apis/event/v1"
+	notifyv1 "github.com/redhat-cop/events-notifier/pkg/apis/notify/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 )
 
@@ -10,4 +12,18 @@ var AddToSchemes runtime.SchemeBuilder
 // AddToScheme adds all Resources to the Scheme
 func AddToScheme(s *runtime.Scheme) error {
 	return AddToSchemes.AddToScheme(s)
+}
+
+type SharedResources struct {
+	Subscriptions map[string]eventv1.EventSubscription
+	Notifiers     map[string]notifyv1.Notifier
+}
+
+func NewSharedResources() SharedResources {
+	es := make(map[string]eventv1.EventSubscription)
+	n := make(map[string]notifyv1.Notifier)
+	return SharedResources{
+		Subscriptions: es,
+		Notifiers:     n,
+	}
 }
