+++ pkg/strings/main.go
@@ -1,0 +1,23 @@
+package strings
+
+//
+// Helper functions to check and remove string from a slice of strings.
+//
+func ContainsString(slice []string, s string) bool {
+	for _, item := range slice {
+		if item == s {
+			return true
+		}
+	}
+	return false
+}
+
+func RemoveString(slice []string, s string) (result []string) {
+	for _, item := range slice {
+		if item == s {
+			continue
+		}
+		result = append(result, item)
+	}
+	return
+}
